Changed size bytes
Decreased the number of bytes that represent the length of a file's name from three to one (16 MB to 256 bytes).
Increased the number of bytes that represent the length of a file's data from three to four (16 MB to 4.2 GB). Note: this value cannot currently be reached due to heap limitations.